# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Go Invoicing API
  version: "1.0.0"
  description: >-
    HTTP API for a multi-tenant invoicing backend. Every product endpoint
    lives under /api/v1; this document's version (1.0.0) tracks that
    contract, not a separate commercial product version. /health and
    /health/db are infrastructure probes and are deliberately excluded from
    the /api/v1 prefix — see the Health tag.


    Tenancy: every protected request's organisation is derived exclusively
    from the authenticated session (the bearer token identifies a user, and
    that user's organisation is fixed at account-creation time). No
    operation accepts a client-supplied organisation identifier, and none
    is documented here. A resource belonging to a different organisation is
    indistinguishable from one that does not exist at all: such a request
    returns 404, never 403, so a client cannot learn anything about another
    tenant's data by probing IDs.


    Authentication: POST /api/v1/auth/login exchanges an email/password for
    an opaque bearer token (not a JWT — no internal structure should be
    assumed or decoded). Send it as `Authorization: Bearer <token>` on every
    protected request. POST /api/v1/auth/logout revokes only the session
    that token belongs to; a user's other active sessions are unaffected.


    Authorization: beyond "authenticated or not", some operations are
    further restricted to specific roles (admin, manager, user). OpenAPI
...


...


...


...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: >-
        Exchanges an email and password for a bearer token. There is no
        organisationId field: the user (and therefore their organisation)
        is resolved from email alone, since email is globally unique.
        Unauthenticated. An unknown email, a wrong password, and an
        inactive account are all indistinguishable to the caller — every
        one of them returns the same generic 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: ada@example.com
              password: correct-horse-battery-staple
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                token: 8f3b1e2c9d4a4f6e8b0a1c2d3e4f5061
                expiresAt: "2026-09-20T09:00:00Z"
                user:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  organisationId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  name: Ada Lovelace
                  email: ada@example.com
                  role: admin
                  isActive: true
                  createdAt: "2026-01-15T10:00:00Z"
                  updatedAt: "2026-01-15T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: { code: unauthorized, message: invalid email or password }
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Log out
      description: >-
        Revokes only the session the presented bearer token resolved to.
        Every other active session belonging to the same user (e.g. from
        another device) is left untouched. Protected.
      responses:
        "204":
          description: The current session was revoked. No body.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/register:
    post:
      tags: [Registration]
      summary: Register a new organisation
      description: >-
        The only way a new organisation and its first user can be created.
        Atomically creates the organisation, its default settings (GBP
        currency, 30-day payment terms, "INV-" invoice prefix), and the
        first user — always with role admin, which cannot be requested or
        overridden by the caller; there is no role field in this request.
        Unauthenticated. This endpoint does not log the new user in: no
        token is issued, and no session is created — call POST
        /api/v1/auth/login afterward to obtain one.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              organisation:
                name: Acme Consulting Ltd
              user:
                name: Ada Lovelace
                email: ada@example.com
                password: correct-horse-battery-staple
      responses:
        "201":
          description: The organisation and its first admin user were created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                organisation:
                  id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  name: Acme Consulting Ltd
                  createdAt: "2026-01-15T10:00:00Z"
                  updatedAt: "2026-01-15T10:00:00Z"
                user:
                  id: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  organisationId: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  name: Ada Lovelace
                  email: ada@example.com
                  role: admin
                  isActive: true
                  createdAt: "2026-01-15T10:00:00Z"
                  updatedAt: "2026-01-15T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: The email is already registered to another user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: { code: email_already_exists, message: a user with this email already exists }
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/organisation:
    get:
      tags: [Organisation]
      summary: Get the caller's organisation
      description: >-
        A self-resource route: always returns the authenticated caller's
        own organisation. There is no path parameter to select a different
        one, so there is no cross-tenant case here. Any authenticated role
        may call this.
      responses:
        "200":
          description: The caller's organisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: The caller's organisation no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Organisation]
      summary: Update the caller's organisation
      description: >-
        Role required: admin only. Partial update: every field is optional
        — an omitted field is left unchanged. An explicitly supplied empty
        string clears that optional field, except name, which may never
        become blank. There is no path parameter; this always targets the
        caller's own organisation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganisationRequest"
            example:
              phone: "+44 20 7946 0958"
              website: https://acme.example.com
      responses:
        "200":
          description: The updated organisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The caller's organisation no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/organisation/settings:
    get:
      tags: [Organisation]
      summary: Get the caller's invoice-numbering settings
      description: >-
        Every authenticated role may read the organisation's current
        invoice currency, payment terms, and invoice-number prefix. The
        internal invoice-number counter itself is never exposed.
      responses:
        "200":
          description: The caller's organisation settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
              example:
                currency: GBP
                paymentTerms: 30
                invoicePrefix: "INV-"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: The caller's organisation has no settings row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags: [Organisation]
      summary: Update the caller's invoice-numbering settings
      description: >-
        Role required: admin only, since these settings affect every future
        invoice the tenant produces. Partial update: an omitted field is
        left unchanged. Unlike PATCH /organisation, none of these three
        fields may be cleared to a blank/zero-ish value by supplying one —
        currency and invoicePrefix are business-required strings, and
        paymentTerms already treats 0 ("due immediately") as a meaningful
        value rather than "unset", so an explicitly supplied blank/invalid
        value is rejected (400), not treated as "clear this field".
        Changing currency here has no effect on any invoice that has
        already been sent — see the Invoices tag for why.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
            example:
              paymentTerms: 14
      responses:
        "200":
          description: The updated settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The caller's organisation has no settings row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create a user
      description: >-
        Role required: admin or manager. An admin may assign any role
        (admin, manager, user); a manager may only create a user with role
        "user" — attempting to assign admin or manager as a manager returns
        403. role is optional and defaults to "user" if omitted. The new
        user always belongs to the caller's own organisation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            example:
              name: Grace Hopper
              email: grace@example.com
              password: another-strong-password
              role: manager
      responses:
        "201":
          description: The created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The caller's organisation no longer exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: A user with this email already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: { code: email_already_exists, message: a user with this email already exists }
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
    get:
      tags: [Users]
      summary: List users
      description: >-
        Role required: admin or manager. Lists users in the caller's own
        organisation. Default sort is createdAt descending (newest team
        member first).
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - name: order
          in: query
          description: Sort direction.
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: sort
          in: query
          description: Field to sort by.
          schema:
            type: string
            enum: [email, role, createdAt]
            default: createdAt
        - name: role
          in: query
          description: Filter to exactly one role.
          schema:
            type: string
            enum: [admin, manager, user]
        - name: active
          in: query
          description: Filter by whether the user is active.
          schema:
            type: boolean
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get a user
      description: >-
        Every authenticated user may retrieve their own record. Retrieving
        another user's record additionally requires the caller to be admin
        or manager — otherwise this returns 403. A user ID belonging to
        another organisation returns 404, the same as one that doesn't
        exist at all.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: The requested user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such user in the caller's organisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: { code: user_not_found, message: user not found }
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/customers/{id}:
    get:
      tags: [Customers]
      summary: Get a customer
      description: Any authenticated role. A customer belonging to another organisation returns 404.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: The requested customer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No such customer in the caller's organisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error: { code: customer_not_found, message: customer not found }
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        An opaque, server-side session token issued by POST
        /api/v1/auth/login — not a JWT, and not decodable by a client.
        Send as `Authorization: Bearer <token>`.
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...